// src/student/model/vo/ProposalState.ts
export enum ProposalState {
  //
  DEFAULT = 'All',
  Drafted = 'Drafted',
  Submitted = 'Submitted',
  Canceled = 'Canceled',
  Approved = 'Approved',
  Rejected = 'Rejected',
  ConsentRequested = 'ConsentRequested',
  ConsentAgreed = 'ConsentAgreed',
  CooperationRequested = 'CooperationRequested',
  CooperationAccepted = 'CooperationAccepted',
  ObservationRequested = 'ObservationRequested',
  ObservationAccepted = 'ObservationAccepted',
}